# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "\/api\/client\/paiement": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Effectuer un paiement",
                "description": "Permet à un client d'effectuer un paiement vers un marchand",
                "operationId": "d040eebc693ade4690eb7e3cdfb21f7f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/PaiementRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paiement effectué avec succès",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#\/components\/schemas\/Transaction"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement effectué avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors du paiement",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "info": {
        "title": "OM Pay API",
        "version": "1.0.0",
        "description": "API de paiement mobile pour OM Pay - Système de paiement et transfert d'argent",
        "contact": {
            "email": "contact@ompay.com"
        }
    },
    "servers": [
        {
            "url": "https:\/\/om-pay-qdx2.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "http:\/\/localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}